+++ packages/most_key_value_editor/lib/src/most_key_value_editor_controller.dart
@@ -1,11 +1,15 @@
 import 'package:flutter/material.dart';
-import 'package:json_diff/json_diff.dart';
 import 'package:json_schema/json_schema.dart' hide ValidationError;
 import 'package:most_schema_parser/most_schema_parser.dart';
 
 import 'json_reader_writer.dart';
-import 'validator/validation_message.dart';
-import 'validator/validation_result.dart';
+import 'most_schema.dart';
+import 'validator/most_form_validator.dart';
+import 'validator/most_unspecified_properties_validator.dart';
+
+typedef ValidatorProvider = MostValidator Function(
+  MostKeyValueEditorController,
+);
 
 /// Editor controller.
 ///
@@ -15,17 +19,33 @@ import 'validator/validation_result.dart';
 /// - storing current json state
 /// - validation
 class MostKeyValueEditorController extends ChangeNotifier {
+  static MostValidator defaultValidator(
+    MostKeyValueEditorController controller,
+  ) {
+    final schema = controller.schema;
+    final jsonRw = controller.jsonRw;
+    final formKey = controller.formKey;
+    return MostMultiValidator([
+      MostJsonSchemaValidator(schema?.jsonSchema),
+      MostFormValidator(formKey),
+      MostUnspecifiedPropertiesValidator(schema?.properties, jsonRw),
+    ]);
+  }
+
   /// Form key that is used to validate Form fields.
   late final GlobalKey<FormState> formKey;
   late final MostJsonSchemaParser _parser;
   late final JsonReaderWriter _jsonRw;
+  late final ValidatorProvider _validatorProvider;
 
   MostSchema? _schema;
   ValidationResult _validationResult = const ValidationResult.notValidated();
 
   MostKeyValueEditorController({
     MostJsonSchemaParser parser = const MostJsonSchemaParser(),
+    ValidatorProvider validatorProvider = defaultValidator,
   })  : _parser = parser,
+        _validatorProvider = validatorProvider,
         formKey = GlobalKey<FormState>() {
     _jsonRw = JsonReaderWriter(
       wrapPropertyRw: _wrapPropertyRw,
@@ -53,6 +73,8 @@ class MostKeyValueEditorController extends ChangeNotifier {
       }
     });
   }
+
+  MostSchema? get schema => _schema;
 
   /// Replaces JSON Schema with the provided one.
   void seedJsonSchema(JsonSchema jsonSchema) {
@@ -103,100 +125,8 @@ class MostKeyValueEditorController extends ChangeNotifier {
   }
 
   ValidationResult _validate(Map<String, dynamic> map) {
-    final schema = _schema;
-
-    List<ValidationMessage> messages = [];
-
-    if (schema == null) {
-      messages.add(
-        const ValidationMessage.error('Schema is not provided'),
-      );
-    } else {
-      final isFormValid = formKey.currentState?.validate() ?? true;
-
-      if (!isFormValid) {
-        messages.add(
-          const ValidationMessage.error('Form contains errors'),
-        );
-      }
-
-      final validationResult = schema.jsonSchema.validate(map);
-
-      if (!validationResult.isValid) {
-        for (final error in [
-          ...validationResult.errors,
-          ...validationResult.warnings
-        ]) {
-          final path = _normalizeSchemaPath(error.instancePath) ?? '';
-          messages.add(
-            ValidationMessage.propertyError(
-              path: path,
-              message: error.message,
-            ),
-          );
-        }
-      }
-
-      final specifiedPropertiesMap = _getMapWithSpecifiedPropertiesOnly(schema);
-      final differ = JsonDiffer.fromJson(specifiedPropertiesMap, map);
-      final diff = differ.diff();
-
-      if (!diff.hasNothing) {
-        messages.add(
-          ValidationMessage.warning(
-            "Source JSON contains unspecified properties:\n"
-            "$diff",
-          ),
-        );
-      }
-    }
-
-    return ValidationResult.validated(messages);
+    final validator = _validatorProvider(this);
+    final validationResult = validator.validate(map);
+    return ValidationResult.validated(validationResult);
   }
-
-  static String? _normalizeSchemaPath(String? path) {
-    if (path == null) return null;
-
-    String normalizedPath = path.replaceAll(' ', '').replaceAll('/', '.');
-    if (normalizedPath.startsWith('.')) {
-      normalizedPath = normalizedPath.substring(1);
-    }
-    return normalizedPath;
-  }
-
-  Map<String, dynamic> _getMapWithSpecifiedPropertiesOnly(MostSchema schema) {
-    final JsonReaderWriter resultRw = JsonReaderWriter();
-
-    void traverseProperties(
-      List<MostProperty> properties, {
-      bool topLevel = false,
-    }) {
-      for (final property in properties) {
-        switch (property) {
-          case MostObjectProperty():
-            if (property.required ||
-                jsonRw.property(property.fullPath).value != null) {
-              traverseProperties(property.properties);
-            }
-          case MostValueProperty():
-            resultRw.property(property.fullPath).value =
-                jsonRw.property(property.fullPath).value;
-        }
-      }
-    }
-
-    traverseProperties(schema.properties, topLevel: true);
-
-    return resultRw.read();
-  }
-}
-
-class MostSchema {
-  final JsonSchema jsonSchema;
-  final List<MostProperty> properties;
-
-  const MostSchema({
-    required this.jsonSchema,
-    required this.properties,
-  });
 }